Regenerate device ID on some versions.
Amplitude is using device ID (instead of user ID) for this scenario:

"The problem is that as part of our
data processing pipeline, we do a bunch of processing of the data to
ensure that the data is "clean" (deduplication, time correction, etc.).
In order to support this, we have designed our platform that all events
from the same device id (which we assume are unique to a specific
device) go to the same event processor.  In this case, where lots of
devices are going to the same event processor, it looks to our platform
like an error and/or attack.  So that device id gets throttled, and the
blocked, and we stop accepting events from it."

in version 1.4.4 of the android library - we weren't sending device id
correctly (we send build id which is *NOT* unique for a phone —
https://github.com/segmentio/analytics-android/blob/d33108eabc0a4132966b7a4b713c3686a482ff9d/core/src/main/java/com/segment/android/info/Device.java#L42).

So we generate a pseudo randomish id on the server and send that to
Amplitude.